refactor: rename `asImagePixelDensitiesSrcSet` file
