+++ src/index.ts
@@ -4,7 +4,7 @@ export { asText } from "./asText";
 export { asHTML } from "./asHTML";
 export { asImageSrc } from "./asImageSrc";
 export { asImageWidthSrcSet } from "./asImageWidthSrcSet";
-export { asImagePixelDensitySrcSet } from "./asImagePixelDensitiesSrcSet";
+export { asImagePixelDensitySrcSet } from "./asImagePixelDensitySrcSet";
 export * as isFilled from "./isFilled";
 
 export { documentToLinkField } from "./documentToLinkField";
